Improved Mushroom Castle + Body Pressure patch
Improved the gameplay around the Mushroom Castle region.
Solved an issue with HP threshold on MoveLifeHandler.
Fixed Body Pressure being refreshed every time it hits a mob.

+++ src/net/server/channel/handlers/MoveLifeHandler.java
@@ -78,8 +78,8 @@ public final class MoveLifeHandler extends AbstractMovementPacketHandler {
 		
 		MobSkill toUse = null;
 		
-		int percHpLeft = (int) ((monster.getHp() / monster.getMaxHp()) * 100);
-
+		int percHpLeft = (int) (((float) monster.getHp() / monster.getMaxHp()) * 100);
+                
 		if (nextMovementCouldBeSkill && monster.getNoSkills() > 0) {
 			int Random = Randomizer.nextInt(monster.getNoSkills());
 			Pair<Integer, Integer> skillToUse = monster.getSkills().get(Random);
